MDEV-21907: InnoDB: Enable -Wconversion on clang and GCC
The -Wconversion in GCC seems to be stricter than in clang.
GCC at least since version 4.4.7 issues truncation warnings for
assignments to bitfields, while clang 10 appears to only issue
warnings when the sizes in bytes rounded to the nearest integer
powers of 2 are different.

Before GCC 10.0.0, -Wconversion required more casts and would not
allow some operations, such as x<<=1 or x+=1 on a data type that
is narrower than int.

The bitwise negation operator appears to promote short integers
to a wider integer type, and hence we must add explicit truncation casts
around them. Microsoft Visual C does not allow a static_cast to
truncate a constant, such as static_cast<uint8_t>(~1). Hence, we will
use the constructor-style cast uint8_t(~1) for such cases.

This has been tested at least with GCC 4.8.5, 5.4.0, 7.4.0, 9.2.1, 10.0.0,
clang 9.0.1, 10.0.0, and MSVC 14.22.27905 (Microsoft Visual Studio 2019).

+++ storage/innobase/fts/fts0que.cc
@@ -606,7 +606,7 @@ fts_ranking_words_add(
 	/* Set ranking words */
 	ut_ad(byte_offset < ranking->words_len);
 	bit_offset = pos % CHAR_BIT;
-	ranking->words[byte_offset] |= 1 << bit_offset;
+	ranking->words[byte_offset] |= static_cast<byte>(1 << bit_offset);
 }
 
 /*******************************************************************//**
